Nessus parser updates:
 - Modularize the skaldship.nessus
 - Added plugin 10399 support (Domain SID user enum)
 - Moved _parse_plugin outside of the for loop
 - Fixed plugin name combinator/slugify
